logical data

+++ dal_facade/DO/Product.cs
@@ -11,7 +11,7 @@ namespace DO
         public int ProductId { get; set; } //Product ID number
         public string ProductName { get; set; } //Product Name
         public int AmmountInStock { get; set; } //Quantity in stock of the product
-        public double ProductPrice { get; set; } //product price
+        public float ProductPrice { get; set; } //product price
         public Category ProductCategory { get; set; }  //Product category
         /// <summary>
         /// Function for printing product details
